src/components/Dashboard.js: moved PieChart data fetch to dashboard
